refactor : menu enum을 새 패키지에 추가, 기존 menu클래스 삭제 및 enum 분할
