+++ src/main/java/subway/view/OutputView.java
@@ -1,7 +1,7 @@
 package subway.view;
 
 import subway.Constant;
-import subway.domain.Menu;
+import subway.domain.menu.MainMenu;
 
 public class OutputView {
 
@@ -11,10 +11,10 @@ public class OutputView {
     public static void printMainView(){
         System.out.print(Constant.VIEW_HEADER);
         System.out.println(Constant.MAIN_VIEW_HEADER);
-        for(Menu.MainMenu menu : Menu.MainMenu.values()) {
+        for(MainMenu menu : MainMenu.values()) {
             System.out.printf(Constant.VIEW_FORMAT, menu.getOrder(), menu.getMessage());
         }
-        System.out.println("");
+        System.out.println();
     }
 
     public static void printManagementView(String menu){
